+++ src/main/kotlin/me/tobiadeyinka/jenkinsci/plugins/builddashboard/job/HealthReport.kt
@@ -1,0 +1,9 @@
+package me.tobiadeyinka.jenkinsci.plugins.builddashboard.job
+
+/**
+ * Health status entity for monitored jobs
+ *
+ * @property[iconUrl] Url for the health status icon
+ * @property[description] Description of the health status
+ */
+class HealthReport constructor(private val iconUrl: String, private val description: String)